move sort_xml_attrs
