+++ lib/iris/__init__.py
@@ -454,3 +454,56 @@ def sample_data_path(*path_to_join):
             "appropriate for general file access.".format(target)
         )
     return target
+
+
+def _sort_xml_attrs(doc):
+    """
+    Takes an xml document and returns a copy with all element
+    attributes sorted in alphabetical order.
+
+    This is a private utility function required by iris to maintain
+    legacy xml behaviour beyond python 3.7.
+
+    Args:
+
+    * doc:
+        The :class:`xml.dom.minidom.Document`.
+
+    Returns:
+        The :class:`xml.dom.minidom.Document` with sorted element
+        attributes.
+
+    """
+    from xml.dom.minidom import Document
+
+    def _walk_nodes(node):
+        """Note: _walk_nodes is called recursively on child elements."""
+
+        # we don't want to copy the children here, so take a shallow copy
+        new_node = node.cloneNode(deep=False)
+
+        # Versions of python <3.8 order attributes in alphabetical order.
+        # Python >=3.8 order attributes in insert order.  For consistent behaviour
+        # across both, we'll go with alphabetical order always.
+        # Remove all the attribute nodes, then add back in alphabetical order.
+        attrs = [
+            new_node.getAttributeNode(attr_name).cloneNode(deep=True)
+            for attr_name in sorted(node.attributes.keys())
+        ]
+        for attr in attrs:
+            new_node.removeAttributeNode(attr)
+        for attr in attrs:
+            new_node.setAttributeNode(attr)
+
+        if node.childNodes:
+            children = [_walk_nodes(x) for x in node.childNodes]
+            for c in children:
+                new_node.appendChild(c)
+
+        return new_node
+
+    nodes = _walk_nodes(doc.documentElement)
+
+    new_doc = Document()
+    new_doc.appendChild(nodes)
+    return new_doc
